+++ dcc-portal-ui/app/scripts/donors/views/donor.html
@@ -294,10 +294,16 @@
                 <th colspan="2" class="text-center"><translate># Donors affected</translate></th>
             </tr>
             <tr>
-                <th class="text-center">
+                <th class="text-center"
+                    data-sortable
+                    data-type="mutations"
+                    data-field="_score"
+                    data-active="true"
+                    data-reversed="true">
                     <abbr data-icgc-tooltip="{{'Breakdown of Affected Donors in [[ projectId ]]<br># Donors affected / # SSM tested donors' | translate | subDelimiters:{projectId: DonorCtrl.donor.projectId} }}">
                         {{ DonorCtrl.donor.projectId }}
-                    </abbr> <i class="icon-caret-down"></i></th>
+                    </abbr>
+                </th>
                 <th class="text-center">
                     <abbr data-icgc-tooltip="{{'# Donors where mutation is observed filtered by current criteria / # SSM tested donors portal wide. Expand to see breakdown by project.' | translate}}"
                           data-tooltip-placement="left"><translate>Across all Projects</translate></abbr></th>
