+++ dcc-portal-ui/app/scripts/index.js
@@ -53,6 +53,10 @@ require('./pathwayviewer/js/controller.js');
 require('./pathwayviewer/js/directive.js');
 require('./keyword/js/keyword.js');
 require('./releases/js/releases.js');
+require('./projects/js/project.js');
+require('./projects/js/project.donor.js');
+require('./projects/js/project.gene.js');
+require('./projects/js/project.mutation.js');
 require('./projects/js/projects.js');
 require('./donors/js/donors.js');
 require('./genes/js/genes.js');
